Pairs Rehab Auto
This is a pairs trading script for those traders who are familiar with the concepts of cointegration and Zscore. If Z-score is below -1.5 it is time to start thinking about a long position that should last a few days or weeks. I have noticed it is most accurate going long. Short plays are extremely risky. People unfamiliar with how these scripts work will lose copious amounts of money so this is only made available for expository purposes, not to actually be used with finances - maybe for paper trading.